e2e: add test for container c/r into pod
This test verifies that container checkpoint and restore into Pod works
when the `container restore --pod` option is used with Pod *name* (this
functionality was previously limited to support only full Pod ID).

Note that crun currently does not support pod restore. This test uses
CRRuntimeSupportsPodCheckpointRestore() and for now it is running only
with runc.

Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>